feat: 홈 화면 조회 추가

+++ src/main/java/com/example/umc10th/domain/mission/entity/Mission.java
@@ -25,7 +25,7 @@ public class Mission extends BaseEntity {
     private Long id;
 
     @Column(name = "mission_point", nullable = false)
-    private Long missionPoint;
+    private Integer missionPoint;
 
     @Column(name = "mission_contents", nullable = false, columnDefinition = "text")
     private String missionContents;
